:art: Normalize bazaar naming and reorganize kernel/bazaar structure
